+++ app/controllers/projects/archives_controller.rb
@@ -1,9 +1,10 @@
 # frozen_string_literal: true
 
 class Projects::ArchivesController < ApplicationController
-  before_action :load_and_authorize_project!
+  include Project::Authorize
 
   def create
+    authorize(@project, :archive?)
     if Project::Archive.new(@project).call
       redirect_to projects_path,
                   notice: "Project archived"
@@ -12,11 +13,4 @@ class Projects::ArchivesController < ApplicationController
       redirect_to project_path(@project)
     end
   end
-
-  private
-
-    def load_and_authorize_project!
-      @project = Project.find(params[:project_id])
-      authorize(@project, :archive?)
-    end
 end